# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Finatrades B2B Commodity Trade Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /b2b/onboarding/company-profile:
    post:
      operationId: submitCompanyProfile
      tags: [onboarding]
      summary: Submit company profile (Step 4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStatus"
  /b2b/onboarding/documents:
    get:
      operationId: getOnboardingDocuments
      tags: [onboarding]
      summary: Get submitted documents list
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnboardingDocument"
    post:
      operationId: submitOnboardingDocuments
      tags: [onboarding]
      summary: Submit KYC/KYB documents (Step 5)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentSubmissionInput"
      responses:
        "200":
          description: Documents submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStatus"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
  /b2b/inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item detail with FUSD valuation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /b2b/inventory/{id}/status:
    patch:
      operationId: updateInventoryStatus
      tags: [inventory]
      summary: Update inventory lifecycle status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /b2b/marketplace/listings/{id}:
    get:
      operationId: getMarketplaceListing
      tags: [marketplace]
      summary: Get listing detail with seller info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /b2b/orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List my orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /b2b/orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /b2b/orders/{id}/payment:
    post:
      operationId: confirmOrderPayment
      tags: [orders]
      summary: Confirm payment for order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfirmInput"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── ESCROW ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OnboardingStatus:
      type: object
      required: [currentStep, totalSteps, completedSteps, kycStatus]
      properties:
        currentStep:
          type: integer
        totalSteps:
          type: integer
        completedSteps:
          type: array
          items:
            type: integer
        kycStatus:
          type: string
          enum: [pending, under_review, approved, rejected]
        approvalDate:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
    CompanyProfileInput:
      type: object
      required: [companyName, registrationNumber, country, address]
      properties:
        companyName:
          type: string
        registrationNumber:
          type: string
        country:
          type: string
        address:
          type: string
        website:
          type: string
        taxId:
          type: string
        businessType:
          type: string
    DocumentSubmissionInput:
      type: object
      required: [documentType, documentUrl]
      properties:
        documentType:
          type: string
          enum: [government_id, proof_of_address, business_registration, tax_id, bank_statement, shareholder_details, beneficial_ownership, authority_letter]
        documentUrl:
          type: string
        notes:
          type: string
    OnboardingDocument:
      type: object
      required: [id, documentType, status, uploadedAt]
      properties:
        id:
          type: string
        documentType:
          type: string
        status:
          type: string
          enum: [pending, verified, rejected]
        uploadedAt:
          type: string
        notes:
          type: ["string", "null"]

    # ── DASHBOARD ──
...
    InventoryStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [available, reserved, pledged, released, sold]
    InventoryItem:
      type: object
      required: [id, inventoryRef, consignmentId, commodityType, quantity, unit, status, warehouseId, fusdValue, createdAt]
      properties:
        id:
          type: string
        inventoryRef:
          type: string
        consignmentId:
          type: string
        ownerId:
          type: string
        ownerName:
          type: string
        commodityType:
          type: string
        grade:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        grossWeight:
          type: ["number", "null"]
        netWeight:
          type: ["number", "null"]
        warehouseId:
          type: string
        warehouseName:
          type: string
        status:
          type: string
          enum: [available, reserved, pledged, released, sold]
        fusdValue:
          type: number
        fusdPerUnit:
          type: number
        inspectionStatus:
          type: string
          enum: [pending, passed, failed]
        moisturePercent:
          type: ["number", "null"]
        qualityGrade:
          type: ["string", "null"]
        warehouseReceiptUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Listing:
      type: object
      required: [id, inventoryId, commodityType, quantity, unit, pricePerUnit, currency, hubId, hubName, sellerId, sellerName, status]
      properties:
        id:
          type: string
        inventoryId:
          type: string
        commodityType:
          type: string
        grade:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        pricePerUnit:
          type: number
        currency:
          type: string
        hubId:
          type: string
        hubName:
          type: string
        hubCountry:
          type: string
        sellerId:
          type: string
        sellerName:
          type: string
        sellerRating:
          type: ["number", "null"]
        incoterms:
          type: string
        status:
          type: string
          enum: [active, reserved, sold, expired]
        certifications:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    OrderInput:
      type: object
      required: [listingId, quantity, fundingMethod]
      properties:
        listingId:
          type: string
        rfqId:
          type: string
        quantity:
          type: number
        fundingMethod:
          type: string
          enum: [bank_transfer, stablecoin, corporate_account, escrow_deposit]
        incoterms:
          type: string
        deliveryWindow:
          type: string
        notes:
          type: string
    PaymentConfirmInput:
      type: object
      required: [paymentMethod, paymentReference]
      properties:
        paymentMethod:
          type: string
        paymentReference:
          type: string
        amountPaid:
          type: number
        currency:
          type: string
    Order:
      type: object
      required: [id, orderRef, buyerId, sellerId, commodity, quantity, unit, totalValue, currency, status, createdAt]
      properties:
        id:
          type: string
        orderRef:
          type: string
        buyerId:
          type: string
        buyerName:
          type: string
        sellerId:
          type: string
        sellerName:
          type: string
        commodity:
          type: string
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        totalValue:
          type: number
        currency:
          type: string
        warehouseId:
          type: string
        warehouseName:
          type: string
        fundingMethod:
          type: ["string", "null"]
        incoterms:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, terms_agreed, payment_pending, payment_confirmed, inventory_reserved, escrow_locked, delivery_in_progress, delivered, completed, cancelled]
        escrowId:
          type: ["string", "null"]
        fusdReference:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    # ── ESCROW ──
...