# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finatrades B2B Commodity Trade Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  orderRef: string;
  buyerId: string;
  buyerName?: string;
  sellerId: string;
  sellerName?: string;
  commodity: string;
  quantity: number;
  unit: string;
  unitPrice?: number;
  totalValue: number;
  currency: string;
  warehouseId?: string;
  warehouseName?: string;
  /** @nullable */
  fundingMethod?: string | null;
  /** @nullable */
  incoterms?: string | null;
  status: OrderStatus;
  /** @nullable */
  escrowId?: string | null;
  /** @nullable */
  fusdReference?: string | null;
  createdAt: string;
  updatedAt?: string;
}